re_view_map
Module visualizers

Re-exports

pub use geo_line_strings::GeoLineStringsVisualizer;

pub use geo_points::GeoPointsVisualizer;

Modules

geo_line_strings 🔒

geo_points 🔒

Structs

GeoSpan

Helper to track an area span in latitude and longitude.

Functions

radius_to_size

Convert a re_types::components::Radius to a re_renderer::Size, considering scene units as meters.

update_span

Extend a span to include another span, if any.